// frontend/src/app/models/medal-front.model.ts
export interface MedalFront {
  id: string;
  name: string;
  description: string;
  type: string;
  size: number;
  width: number;
  height: number;
  backgroundColor: string;
  logoColor: string;
  logoSize: number;
  logoX: number;
  logoY: number;
  borderRadius: number;
  useBackgroundImage: boolean;
  createdAt: string;
  updatedAt: string;
}